EXPERIMENTAL: Experimental allocator improvement
A substitutional plan for now.

The following is the original patch description of Hans:
----------------------------------------------------------------
Change the zone allocator to prioritize filling up zones.
This leads to higher space amplification (~2x) but reduces
write long-tail latencies greatly by reducing zone-finish
operations.

The better solution to the tail latency problems with
finish is to finish zones out of the allocator loop as
part of a background job, so that is what we should do in stead.

Signed-off-by: Changlong Chen <chenchanglong@bytedance.com>